requirements.txt
# Core LiveKit Framework
livekit==1.0.19
livekit-agents>=1.3.6
livekit-api==1.0.7
livekit-blingfire==1.0.0
livekit-protocol==1.1.0

# LiveKit Plugins
livekit-plugins-openai==1.3.3
livekit-plugins-speechmatics==1.3.3
livekit-plugins-silero==1.3.3
livekit-plugins-turn-detector==1.3.3
livekit-murf>=0.1.0

# OpenTelemetry (required by livekit-agents, fixes compatibility issues)
opentelemetry-api>=1.39.0
opentelemetry-sdk>=1.39.0
opentelemetry-exporter-otlp>=1.39.0

# OpenAI Client
openai>=2.8.1

# Environment Variables
python-dotenv>=1.2.1

# Audio Input
sounddevice==0.5.3

# Speechmatics Runtime
speechmatics-rt==0.5.1
